// src/renderer/features/visualizer/components/audiomotionanalyzer/visualizer-settings-options.d.ts
export declare const modeOptions: {
    label: string;
    value: string;
}[];
export declare const colorModeOptions: {
    label: string;
    value: string;
}[];
export declare const gradientOptions: {
    label: string;
    value: string;
}[];
export declare const channelLayoutOptions: {
    label: string;
    value: string;
}[];
export declare const fftSizeOptions: {
    label: string;
    value: string;
}[];
export declare const frequencyScaleOptions: {
    label: string;
    value: string;
}[];
export declare const weightingFilterOptions: {
    label: string;
    value: string;
}[];
export declare const minFreqOptions: {
    label: string;
    value: string;
}[];
export declare const maxFreqOptions: {
    label: string;
    value: string;
}[];
export declare const barSpaceOptions: {
    label: string;
    value: string;
}[];